linux: Round font sizes when going from points to pixels.
Round point-based Pango font sizes to the nearest pixel.
Otherwise, we can end up with smaller UI fonts than GTK+ 2.0
apps in some cases.

This seems to be in line with what get_scaled_size() does in
Pango's pango/pangofc-fontmap.c, and gives me identically
sized text in Gimp 2.6.12 and Chrome when I set the
Gtk/FontName XSETTINGS property to "Arial 10".

Note that Chrome doesn't currently honor additional
fine-grained configuration performed via FontConfig, like
disabling antialiasing for fonts of a certain size. That'll
happen in a separate change.

BUG=375824

Review URL: https://codereview.chromium.org/377413002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@282204 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: dd06564986bdc86b90e6ea2a87dbadb019da2345

+++ ui/gfx/pango_util.h
@@ -52,7 +52,7 @@ void SetUpPangoLayout(
     int flags);
 
 // Returns the size in pixels for the specified |pango_font|.
-size_t GetPangoFontSizeInPixels(PangoFontDescription* pango_font);
+int GetPangoFontSizeInPixels(PangoFontDescription* pango_font);
 
 // Retrieves the Pango metrics for a Pango font description. Caches the metrics
 // and never frees them. The metrics objects are relatively small and very
